# nodejs-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is the API for the Estuary application."
  version: "0.0.0"
  title: "Estuary API"
  termsOfService: "http://estuary.tech"
  contact:
    name: "API Support"
    url: "https://docs.estuary.tech/feedback"
  license:
    name: "Apache 2.0 Apache-2.0 OR MIT"
    url: "https://github.com/application-research/estuary/blob/master/LICENSE.md"
host: "api.estuary.tech"
basePath: "/"
security:
- bearerAuth: []
paths:
...
  /public/by-cid/{cid}:
    get:
      tags:
      - "public"
      summary: "Get Content by Cid"
      description: "This endpoint returns the content associated with a CID"
      operationId: "publicBy_cidCidGET"
      produces:
      - "application/json"
      parameters:
      - name: "cid"
        in: "path"
        description: "Cid"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Public"
...
  /public/miners/deals/{miner}:
    get:
      tags:
      - "public"
      - "miner"
      summary: "Get all miners deals"
      description: "This endpoint returns all miners deals"
      operationId: "publicMinersDealsMinerGET"
      produces:
      - "application/json"
      parameters:
      - name: "miner"
        in: "path"
        description: "Filter by miner"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Public"
  /public/miners/failures/{miner}:
    get:
      tags:
      - "public"
      - "net"
      summary: "Get all miners"
      description: "This endpoint returns all miners"
      operationId: "publicMinersFailuresMinerGET"
      produces:
      - "application/json"
      parameters:
      - name: "miner"
        in: "path"
        description: "Filter by miner"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Public"
  /public/miners/stats/{miner}:
    get:
      tags:
      - "public"
      - "miner"
      summary: "Get miner stats"
      description: "This endpoint returns miner stats"
      operationId: "publicMinersStatsMinerGET"
      produces:
      - "application/json"
      parameters:
      - name: "miner"
        in: "path"
        description: "Filter by miner"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Public"
  /public/miners/storage/query/{miner}:
    get:
      tags:
      - "deals"
      summary: "Query Ask"
      description: "This endpoint returns the ask for a given CID"
      operationId: "publicMinersStorageQueryMinerGET"
      produces:
      - "application/json"
      parameters:
      - name: "miner"
        in: "path"
        description: "CID"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Deals"
...
securityDefinitions:
  bearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"
...